+++ postgres_searchindex/conf.py
@@ -1,5 +1,6 @@
 from django.conf import settings
 
+# index config
 POSTGRES_SEARCHINDEX = getattr(
     settings,
     "POSTGRES_SEARCHINDEX",
@@ -8,6 +9,12 @@ POSTGRES_SEARCHINDEX = getattr(
     },
 )
 
+# search view
+QUERY_FUNC = getattr(
+    settings,
+    "POSTGRES_SEARCHINDEX_QUERY_FUNC",
+    "postgres_searchindex.query_helpers.basic_search_with_ranking",
+)
 LANGUAGE_2_PGCONFIG = getattr(
     settings,
     "POSTGRES_SEARCHINDEX_LANGUAGE_2_PGCONFIG",
@@ -17,7 +24,13 @@ LANGUAGE_2_PGCONFIG = getattr(
         "fr": "french",
     },
 )
+PAGINATE_BY = getattr(
+    settings,
+    "POSTGRES_SEARCHINDEX_PAGINATE_BY",
+    20,
+)
 
+# contrib.djangocms
 USE_CMS_INDEX = getattr(
     settings,
     "POSTGRES_SEARCHINDEX_USE_CMS_INDEX",