+++ postgres_searchindex/management/commands/postgres_searchindex_update.py
@@ -1,10 +1,10 @@
 from django.core.management import BaseCommand
 
-from postgres_searchindex.management.indexing import update_indexes
+from postgres_searchindex.management.indexing import update_index
 
 
 class Command(BaseCommand):
     help = "update index"
 
     def handle(self, *args, **options):
-        update_indexes()
+        update_index(self.stdout)