+++ CHANGELOG.md
@@ -1,1 +1,6 @@
 === (ongoing) ===
+=== 0.0.2 ===
+
+- better site support
+- enhance cms contrib
+- use built in views with your customized search function